implement RequestOptionsBase, RestReponse, strip methods
