+++ libraries/botframework-connector/src/utils/index.ts
@@ -1,2 +1,1 @@
-export * from './serviceCallback';
-export * from './userAgent';
+export * from './httpUtils'
